Fix front cart bug, add red cart sprite
A typo resulted in the front cart determination logic not working. This
is now behaving correctly.

Add a simple red cart sprite as an experiment. The player object
transform no longer has a 2x horizontal scale, so child carts that
aren't using this sprite now appear as squares. This will be remedies as
additional sprites are added.