Add blue cart sprite, small bug fix
Add a basic sprite for a blue car and use this for the ParkedVehicle
prefabs. Note that the vehicles are now larger than before as a result.
Also fix a small bug where magnetism would remain requested after
running out of time, re-enabling the powerup once a red cart was
returned.