Add magnetism error sound, don't apply effect on spindown
Add an error sound that plays when the player requests magnetism when no
magnetism time is available. Also adjust the behavior such that the
actual magnetism force effect is not applied during spindown. This makes
the spindown an actual penalty that encourages the user to not rapidly
request an un-request magnetism.