First pass at UI Icons and Animations
Add icons for dashes, magnetism, and speed. Update scripts to animate
them: dashes are still displayed as a text number, but magnetism time is
now represented by a bar, and speed is represented by a needle on a
speedometer.

Increase speed update rate to 10 Hz instead of 1 Hz to make the change
more gradual.

Re-arrange all UI elements to the bottom of the sceen. Add a translucent
bar to help provide contrast from white text in the interface. However,
this currently blocks a portion of the playable area; this needs to be
updated.