Minor UI icon modifications, adjust scale and position
Added new UI icons that increase attention on the area that conveys
information to the player (the text for dashes, and the bar for
magnetism). Slightly shrink the icons and move everything to a tighter
position at the bottom of the screen.

Still need to adjust the bottom of the playable area.